optimizador_cvs/requirements.txt
streamlit
python-dotenv
fpdf2
langchain-community
langchain-google-genai
pypdf
